Move code that does the file conversion into pyrs.core
